Improve algorithm for jumping out of delimiters
The "jump out" feature depends mostly on the function
`pear_tree#GetDelimiterAfterCursor()` which in turn depends on
`pear_tree#PairsInText()`.

Previously, the work of determining if the cursor was at a delimiter
was mostly left to the latter function, which would (in short) return
column numbers that were determined to be at the start of a delimiter.

The work is now more evenly divided between the two functions.
`pear_tree#PairsInText()` gathers every opener-delimiter pair in a given
line of text, and `pear_tree#GetDelimiterAfterCursor()` uses this result
to check if the text after the cursor is a balanced delimiter.

As a result, `pear_tree#PairsInText()` is faster and less complex, and
`pear_tree#GetDelimiterAfterCursor()` is much more accurate.