Improve performance of `Traverser.TraverseText()`
Rather than attempting to step the traverser down the trie at every
character in the text, the changes use the fact that, more often than
not, _most_ characters in the text are not in the trie. So instead of
checking every character in the text, use `stridx()` to skip forward to
characters that are in the trie. This method is much faster, taking
under one second to traverse a ~60,000-character line of lisp code.

Additionally, checking if a string or list is empty is now done by
direct comparison to '' or [] respectively; this is slightly faster
than checking if their length is 0 using `len()` or `strlen()`.