Improve insert mode performance on long lines
The trie traverser needs to traverse the current line when the cursor
moves in insert mode. Rather than scan the entire line, the traverser
can for every string in the trie find the last index at which the
string's final character appeared. Traversal may then start at the
earliest of these indexes, often drastically reducing the amount of
text that must be processed.