Fix issue causing slow insert mode movement
This prevents TraverseBuffer from considering a potentially large number
of multicharacter openers in the buffer. Instead it will consider
wildcard openers beyond the calculated starting point. Then, when no
more are found, it will skip to the end of the buffer minus the length
of the longest opener in the trie and proceed with a simple linear scan.