Don't rescan buffer when moving backwards/linewise
Though buffer scanning is fairly fast, it causes noticeable lag when
moving up or down rapidly through larger files. Instead of scanning
the buffer on every such cursor movement, set a flag telling the
traverser to rescan the buffer before a character is inserted or the
cursor is moved forward along a single line.