Fix: Try to step from root after traverser reset
Problem: If {"f'": "'"} is a pair, and the user types ff', the closing '
is not inserted. After the second f is inserted, the traverser resets to
the trie root but fails to step to the f node like it should.

Solution: After the traverser is reset because the user typed a
character that is not the next character of the opening string,
immediately try to use that character as the start of a different
opening string.